+++ libftprintf/integer/ftprt_put_unumber.c
@@ -6,7 +6,7 @@ static void ftprt_put_unumber_apo(uintmax_t val, size_t len, t_printff *fl,
     uintmax_t   pw;
     size_t      delim_c;
 
-    if (fl->precision != -1 && (fl->precision > len))
+    if (fl->precision != -1 && (fl->precision > (long long)len))
         ftprt_putnchar('0', fl->precision - len, f_putchar);
     if (val || fl->precision != 0)
     {
@@ -32,7 +32,7 @@ void        ftprt_put_unumber(uintmax_t val, size_t len, t_printff *fl,
 
     if (fl->flags[4])
         return (ftprt_put_unumber_apo(val, len, fl, f_putchar));
-	if (fl->precision != -1 && (fl->precision > len))
+	if (fl->precision != -1 && (fl->precision > (long long)len))
 		ftprt_putnchar('0', fl->precision - len, f_putchar);
 	if (val || fl->precision != 0)
 	{
